refactor: move app to /app, marketing page stub
