+++ src/routes/+page.svelte
@@ -1,15 +1,9 @@
-<script lang="ts">
-  import { goto } from '$app/navigation';
-  import wallet from '$lib/stores/wallet';
+<p>Marketing page goes here</p>
 
-  $: {
-    if ($wallet.connected) goto('/dashboard');
+<a href="/app/" class="typo-link">Go to the app</a>
+
+<style>
+  p {
+    color: var(--color-positive);
   }
-</script>
-
-<h1>Welcome page</h1>
-<p>
-  This page will urge the user to connect their wallet if they haven't yet, wait for all the stores
-  to be ready and then forward to the dashboard
-</p>
-<a href="/component-showcase" sveltekit:prefetch>Go to component showcase</a>
+</style>
