+++ src/routes/app/component-showcase/examples/example-stepper-steps/step-2.svelte
