+++ src/routes/app/component-showcase/examples/example-stepper-steps/success-step.svelte
