+++ src/routes/app/dashboard/sections/create-stream-flow/input-details.svelte
